+++ tests/unit/graph/graph-spec.ts
@@ -1158,9 +1158,9 @@ describe('mapper fn', () => {
     graph.setItemState(node, 'custom', true);
     expect(keyShape.attr('green'));
 
+    // clear all states of  the item
     graph.clearItemStates(node);
-    // green
-    expect(keyShape.attr('fill')).toEqual('green');
+    expect(keyShape.attr('fill')).toEqual('#666');
 
     const edge = graph.addItem('edge', { id: 'edge2', source: 'node', target: 'node2Mapped' });